requirements.txt
fastapi
uvicorn
python-multipart
passlib[bcrypt]
python-jose
python-dotenv
google-generativeai
jinja2
PyPDF2
